refactor(connector): per-dialect types in types_<dialect>.go
Moves postgresSpec and snowflakeSpec out of types.go into dedicated
types_postgres.go and types_snowflake.go. types.go keeps only shared
wire shapes (createReq, updateReq, configEnvelope, createResp,
getConnectorResp). Scales cleanly as new dialects land.
